Move STM Unique ID function/callback to stm_flash
